libexpr: Use 16-byte atomic loads/stores MOVAPS/MOVDQA for ValueStorage<8>
This would have to pair with CMPXCHG16B (or other double-width CAS approaches)
to be able to implement concurrent thunk evaluation.

See the extensive comment in the code for the reasons why we have to
do it this way. The motivation for the change is to simplify the eventual
implementation of parallel evaluation. Being able to atomically update the
whole Value is much easier to reason about.